양방향 참조 삭제 오류 수정 / 로그인 여부 확인 api 추가

+++ bookcafe/src/main/java/com/miniproject2/bookcafe/controller/MoimController.java
@@ -1,10 +1,12 @@
 package com.miniproject2.bookcafe.controller;
 
+import com.miniproject2.bookcafe.domain.Comment;
 import com.miniproject2.bookcafe.domain.Moim;
 import com.miniproject2.bookcafe.domain.MoimMember;
 import com.miniproject2.bookcafe.dto.MoimMemberRequestDto;
 import com.miniproject2.bookcafe.dto.MoimRequestDto;
 import com.miniproject2.bookcafe.dto.MoimResponseDto;
+import com.miniproject2.bookcafe.repository.CommentRepository;
 import com.miniproject2.bookcafe.repository.MoimMemberRepository;
 import com.miniproject2.bookcafe.repository.MoimRepository;
 import com.miniproject2.bookcafe.service.MoimService;
@@ -20,6 +22,7 @@ public class MoimController {
     private final MoimRepository moimRepository;
     private final MoimService moimService;
     private final MoimMemberRepository moimMemberRepository;
+    private final CommentRepository commentRepository;
 
     // 모임 생성하기
     @PostMapping("/moims")
@@ -31,6 +34,9 @@ public class MoimController {
     // 모임 삭제하기
     @DeleteMapping("/moims/{moimId}")
     public Long deletMoim(@PathVariable Long moimId){
+        Moim moim =  moimRepository.findById(moimId).orElseThrow(
+                () -> new IllegalArgumentException("모임이 존재하지 않습니다. ")
+        );
         moimRepository.deleteById(moimId);
         return moimId;
     }
@@ -70,10 +76,15 @@ public class MoimController {
         MoimMember moimMember =
                 moimMemberRepository.findByMoimIdAndNickname(
                         requestDto.getMoimId(), requestDto.getNickname());
+        if(moimMember == null){
+            throw new IllegalArgumentException("모임에 참가하지 않았습니다.");
+        }
+
         Long memberId = moimMember.getMemberId();
         moimMemberRepository.deleteById(memberId);
         return null;
     }
+
 }
